+++ oblate/fields/structs.py
@@ -25,6 +25,8 @@ from __future__ import annotations
 from typing import (
     TYPE_CHECKING,
     Dict as DictT,
+    List as ListT,
+    Set as SetT,
     TypedDict as TypedDictT,
     Any,
     Union,
@@ -41,6 +43,8 @@ if TYPE_CHECKING:
 __all__ = (
     'Dict',
     'TypedDict',
+    'List',
+    'Set',
 )
 
 TD = TypeVar('TD', bound=TypedDictT)
@@ -49,7 +53,28 @@ VT = TypeVar('VT')
 MISSING = utils.MISSING
 
 
-class Dict(Field[DictT[KT, VT], DictT[KT, VT]]):
+class _BaseStructField(Field[KT, VT]):
+    _struct_name: str
+    _friendly_struct_name: str
+
+    ERR_INVALID_DATATYPE = 'struct.invalid_datatype'
+    ERR_TYPE_VALIDATION_FAILED = 'struct.type_validation_failed'
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        self.ERR_INVALID_DATATYPE = f'{self._struct_name}.invalid_datatype'  # type: ignore
+        self.ERR_TYPE_VALIDATION_FAILED = f'{self._struct_name}.type_validation_failed'  # type: ignore
+        super().__init__(*args, **kwargs)
+
+    def format_error(self, error_code: Any, context: ErrorContext) -> Union[FieldError, str]:
+        if error_code == self.ERR_INVALID_DATATYPE:
+            return f'Value must be a {self._friendly_struct_name}'
+        if error_code == self.ERR_TYPE_VALIDATION_FAILED:
+            return FieldError(context.metadata['type_validation_fail_errors'])
+
+        return super().format_error(error_code, context)  # pragma: no cover
+
+
+class Dict(_BaseStructField[DictT[KT, VT], DictT[KT, VT]]):
     """A field that accepts a dictionary.
 
     When initialized without an argument, the field accepts any arbitrary
@@ -77,10 +102,8 @@ class Dict(Field[DictT[KT, VT], DictT[KT, VT]]):
     value_tp:
         The type of value of dictionary.
     """
-    __slots__ = ('_tp',)
-
-    ERR_INVALID_DATATYPE = 'dict.invalid_datatype'
-    ERR_TYPE_VALIDATION_FAILED = 'dict.type_validation_failed'
+    _struct_name = 'dict'
+    _friendly_struct_name = 'dictionary'
 
     def __init__(self, key_tp: Type[KT] = MISSING, value_tp: Type[VT] = MISSING, /, **kwargs: Any) -> None:
         if key_tp is not MISSING:
@@ -91,14 +114,6 @@ class Dict(Field[DictT[KT, VT], DictT[KT, VT]]):
             self._tp = None
 
         super().__init__(**kwargs)
-
-    def format_error(self, error_code: Any, context: ErrorContext) -> Union[FieldError, str]:
-        if error_code == self.ERR_INVALID_DATATYPE:
-            return 'Value must be a dictionary'
-        if error_code == self.ERR_TYPE_VALIDATION_FAILED:
-            return FieldError(context.metadata['type_validation_fail_errors'])
-
-        return super().format_error(error_code, context)  # pragma: no cover
 
     def value_load(self, value: Any, context: LoadContext) -> DictT[KT, VT]:
         if not isinstance(value, dict):
@@ -114,7 +129,7 @@ class Dict(Field[DictT[KT, VT], DictT[KT, VT]]):
         return value
 
 
-class TypedDict(Field[TD, TD]):
+class TypedDict(_BaseStructField[TD, TD]):
     """A field that validates from a :class:`typing.TypedDict`.
 
     This class provides type validation on the values of given :class:`TypedDict`,
@@ -134,20 +149,12 @@ class TypedDict(Field[TD, TD]):
     typed_dict: :class:`typing.TypedDict`
         The typed dictionary to validate from.
     """
-    ERR_INVALID_DATATYPE = 'typed_dict.invalid_datatype'
-    ERR_TYPE_VALIDATION_FAILED = 'typed_dict.type_validation_failed'
+    _struct_name = 'typed_dict'
+    _friendly_struct_name = 'dictionary'
 
     def __init__(self, typed_dict: Type[TD], /, **kwargs: Any):
         self._typed_dict = typed_dict
         super().__init__(**kwargs)
-
-    def format_error(self, error_code: Any, context: ErrorContext) -> Union[FieldError, str]:
-        if error_code == self.ERR_INVALID_DATATYPE:
-            return 'Value must be a dictionary'
-        if error_code == self.ERR_TYPE_VALIDATION_FAILED:
-            return FieldError(context.metadata['type_validation_fail_errors'])
-
-        return super().format_error(error_code, context)  # pragma: no cover
 
     def value_load(self, value: Any, context: LoadContext) -> TD:
         if not isinstance(value, dict):
@@ -161,4 +168,64 @@ class TypedDict(Field[TD, TD]):
         return value  # type: ignore
 
     def value_dump(self, value: TD, context: DumpContext) -> TD:
+        return value
+
+
+class List(_BaseStructField[ListT[KT], ListT[KT]]):
+    """A field that accepts a :class:`list`.
+
+    Parameters
+    ----------
+    tp:
+        The type of list elements. This parameter corresponds to ``T``
+        in ``typing.List[T]``.
+    """
+    _struct_name = 'list'
+    _friendly_struct_name = 'list'
+
+    def __init__(self, tp: Type[KT] = MISSING, /, **kwargs: Any):
+        self._tp = None if tp is MISSING else ListT[tp]
+        super().__init__(**kwargs)
+
+    def value_load(self, value: Any, context: LoadContext) -> ListT[KT]:
+        if not isinstance(value, list):
+            raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
+        if self._tp is not None:  # type: ignore
+            validated, errors = utils.validate_struct(value, self._tp, stack_errors=True)  # type: ignore
+            if not validated:
+                metadata = {'type_validation_fail_errors': errors}
+                raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)
+        return value  # type: ignore
+
+    def value_dump(self, value: ListT[KT], context: DumpContext) -> ListT[KT]:
+        return value
+
+
+class Set(_BaseStructField[SetT[KT], SetT[KT]]):
+    """A field that accepts a :class:`set`.
+
+    Parameters
+    ----------
+    tp:
+        The type of set elements. This parameter corresponds to ``T``
+        in ``typing.Set[T]``.
+    """
+    _struct_name = 'set'
+    _friendly_struct_name = 'set'
+
+    def __init__(self, tp: Type[KT] = MISSING, /, **kwargs: Any):
+        self._tp = None if tp is MISSING else SetT[tp]
+        super().__init__(**kwargs)
+
+    def value_load(self, value: Any, context: LoadContext) -> SetT[KT]:
+        if not isinstance(value, set):
+            raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
+        if self._tp is not None:  # type: ignore
+            validated, errors = utils.validate_struct(value, self._tp, stack_errors=True)  # type: ignore
+            if not validated:
+                metadata = {'type_validation_fail_errors': errors}
+                raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)
+        return value  # type: ignore
+
+    def value_dump(self, value: SetT[KT], context: DumpContext) -> SetT[KT]:
         return value